requirements.txt
aiogram==3.6.0
APScheduler==3.10.4
python-dotenv==1.0.1